+++ packages/orq-rc/RELEASES.md
@@ -9679,3 +9679,13 @@ Based on:
 - [typescript v4.10.0-rc.72] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.72] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.72 - packages/orq-rc
+
+## 2026-06-02 14:48:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.770.0 (2.893.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.73] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.73] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.73 - packages/orq-rc